Update perlin3d.py
Fixed bug
line 35, in generate_perlin_noise_3d
    theta = 2*np.pi*np.random.rand(res[0] + 1, res[1] + 1, res[2] + 1)
  File "numpy/random/mtrand.pyx", line 1219, in numpy.random.mtrand.RandomState.rand
  File "numpy/random/mtrand.pyx", line 437, in numpy.random.mtrand.RandomState.random_sample
  File "_common.pyx", line 307, in numpy.random._common.double_fill
TypeError: 'float' object cannot be interpreted as an integer

+++ perlin_numpy/perlin3d.py
@@ -107,7 +107,8 @@ def generate_fractal_noise_3d(
     for _ in range(octaves):
         noise += amplitude * generate_perlin_noise_3d(
             shape,
-            (frequency*res[0], frequency*res[1], frequency*res[2]),
+            # (frequency*res[0], frequency*res[1], frequency*res[2]),
+            (int(frequency * res[0]), int(frequency * res[1]), int(frequency * res[2])),
             tileable,
             interpolant
         )
